+++ CHANGELOG.md
@@ -15,8 +15,10 @@ We follow the [Semantic Versioning 2.0.0](http://semver.org/) format.
 - Updates loan origination fee displays to work with corrected constants
 - Fix alignment issues in debt summary
 - Update our national statistics from scorecard project
+- Add form-level error for bad program ID
 - Update national stats to deliver generic stats at /api/national-stats/
 - simplify warnings for ID problems
+- make bare /offer/ URL return blank disclosure
 
 ## 2.1.5
 - Show recalculation updates on mobile screens